Move SVG chart generator to server-kotlin module
