+++ server-kotlin/build.gradle
@@ -38,6 +38,11 @@ dependencies {
     compile "io.ktor:ktor-auth-jwt:$ktor_version"
     compile "io.ktor:ktor-jackson:$ktor_version"
     testCompile "io.ktor:ktor-server-tests:$ktor_version"
+    compile("io.ktor:ktor-client-json:$ktor_version")
+    compile "io.ktor:ktor-client-gson:$ktor_version"
+    compile "io.ktor:ktor-client-core:$ktor_version"
+    compile "io.ktor:ktor-client-core-jvm:$ktor_version"
+    compile "io.ktor:ktor-client-apache:$ktor_version"
 
     // google docs API
     compile 'com.google.api-client:google-api-client:1.30.1'
@@ -51,6 +56,12 @@ dependencies {
 
     //Akka actor model
     compile group: 'com.typesafe.akka', name: 'akka-actor_2.13', version: '2.6.0'
+
+    // SVG DSL
+    compile "com.github.nwillc:ksvg:2.3.3"
+
+
+
 }
 
 jar {
