+++ server-kotlin/src/chart_generator/pojo/Budget.kt
@@ -1,0 +1,15 @@
+package chart_generator.pojo
+
+import java.math.BigDecimal
+
+data class Etl(val _id: String, val vality: String)
+
+data class Budget(
+        val _id: String,
+        val year: Int,
+        val etl: Etl,
+        val budgetExpenditureAmount: BigDecimal,
+        val expenditureAmount: BigDecimal,
+        val budgetIncomeAmount: BigDecimal,
+        val incomeAmount: BigDecimal
+)
