--- svg_char_generator/gradle.properties
@@ -1,4 +1,0 @@
-ktor_version=1.2.6
-kotlin.code.style=official
-kotlin_version=1.3.60
-logback_version=1.2.1
