--- svg_char_generator/resources/application.conf
@@ -1,11 +1,0 @@
-ktor {
-    deployment {
-        port = 8080
-        port = ${?PORT}
-        watch = [ svg_char_generator ]
-        autoreload = true
-    }
-    application {
-        modules = [ cz.cityvizor.svg_char_generator.ApplicationKt.module ]
-    }
-}
