--- svg_char_generator/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = "svg_char_generator"
